# swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IRef": {
				"properties": {
					"name": {
						"type": "string",
						"description": "Имя"
					},
					"version": {
						"type": "number",
						"format": "double",
						"description": "Версия справочника"
					},
					"lastUpdate": {
						"type": "string",
						"format": "date-time",
						"description": "Время последнего обновления\r\nФормат: Date Long"
					}
				},
				"required": [
					"name",
					"version",
					"lastUpdate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IRightInfoMeta": {
				"properties": {
					"ref": {
						"$ref": "#/components/schemas/IRef"
					}
				},
				"required": [
					"ref"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserRights": {
				"enum": [
					0,
					1,
					2,
					3,
					4,
					5,
					6,
					7,
					8,
					9,
					10,
					11,
					12,
					13,
					14,
					15,
					16,
					17,
					18,
					19,
					20,
					21,
					22,
					23,
					24,
					25,
					26,
					27,
					28,
					29,
					30,
					31,
					32,
					33,
					34,
					35,
					36,
					37,
					38,
					39,
					40,
					41,
					42,
					43,
					44,
					45,
					46,
					47,
					48,
					49,
					50,
					51,
					52,
					53,
					54,
					55,
					56,
					57,
					58,
					59,
					60,
					61,
					62,
					63,
					64,
					65,
					66,
					67,
					68,
					69,
					70,
					71,
					72
				],
				"type": "number"
			},
			"IRightsResponse": {
				"properties": {
					"meta": {
						"$ref": "#/components/schemas/IRightInfoMeta"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/UserRights"
						},
						"type": "array"
					},
					"error": {
						"items": {
							"properties": {
								"message": {
									"type": "string"
								},
								"code": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"message",
								"code"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IInfoMeta": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"IInfoData": {
				"properties": {
					"serverName": {
						"type": "string"
					},
					"versionName": {
						"type": "string"
					},
					"versionCode": {
						"type": "number",
						"format": "double"
					},
					"version": {
						"type": "string"
					},
					"availableRights": {
						"items": {
							"$ref": "#/components/schemas/UserRights"
						},
						"type": "array"
					}
				},
				"required": [
					"serverName",
					"versionName",
					"versionCode",
					"version",
					"availableRights"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IInfoResponse": {
				"properties": {
					"meta": {
						"$ref": "#/components/schemas/IInfoMeta"
					},
					"data": {
						"$ref": "#/components/schemas/IInfoData"
					},
					"error": {
						"items": {
							"properties": {
								"message": {
									"type": "string"
								},
								"code": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"message",
								"code"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"clientAccessToken": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "tornado-etl-evo-server",
		"version": "1.0.0",
		"license": {
			"name": "Commercial"
		},
		"contact": {
			"name": "Eugene Grebennikov",
			"email": "djonnyx@gmail.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/rights": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IRightsResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"meta": {
												"ref": {
													"name": "rights",
													"version": 1,
													"lastUpdate": "2021-09-15T11:04:49.030Z"
												}
											},
											"data": [
												0,
												1,
												2,
												3,
												4,
												5,
												6,
												7,
												8,
												9,
												10,
												11,
												12,
												13,
												14,
												15,
												16,
												17,
												18,
												19,
												20,
												21,
												22,
												23,
												24,
												25,
												26,
												27,
												28,
												34,
												35,
												36,
												37,
												38,
												39,
												40,
												41,
												42,
												43,
												44,
												45,
												46,
												47,
												48,
												54,
												55,
												56,
												57,
												58,
												59,
												60,
												61,
												62,
												63,
												64,
												65,
												66,
												67,
												68,
												69,
												70,
												71,
												72
											]
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Rights"
				],
				"security": [
					{
						"clientAccessToken": []
					}
				],
				"parameters": []
			}
		},
		"/Info": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IInfoResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"meta": {},
											"data": {
												"serverName": "Evotor",
												"versionName": "Evo-5X"
											}
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Info"
				],
				"security": [
					{
						"clientAccessToken": []
					}
				],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}